fix expression parsing infinite loop
IIRC expressions used to be required and so a call to `expression()`
implied an error if none was found. At some point, expression() became
optional. Actually, it was probably for implementing adjoining function
arguments. Anyway, we forgot to account for this lack of error in
existing loops, etc. Statements beginning with an operator caused
infinite loops.